let change log markdown has it's own class.
